Release v0.4.10: IPv6 fragment reassembly (SNB-0014) (#98)
* fix(capture): reassemble IPv6-fragmented packets (SNB-0014)

The IPv4-fragment fix (SNB-0011) handled IPv4 only. IPv6 carries fragmentation
in a Fragment extension header, not the base header, but parse_packet hardcoded
ip_id/fragment_offset/more_fragments to None/None/false for IPv6 — so
PacketProcessor never routed IPv6 fragments to the reassembler and each fragment
was mis-parsed as a whole datagram, silently dropping the SIP message (a large
INVITE with SDP over IPv6 decoded to 0; tshark saw it).

parse_packet now walks the IPv6 extension headers and, on a Fragment header,
pulls its offset / MF / 32-bit identification so the existing reassembler keys
and reassembles it (reparse_transport then recovers the UDP/TCP header from the
reassembled payload). ParsedPacket.ip_id widened u16 -> u32 (the IPv6 fragment
id is 32-bit; the IPv4 16-bit id casts up).

Verified: a fragment6()-split IPv6 OPTIONS now decodes; capture-module tests
green after the ip_id widening.

* Release v0.4.10: IPv6 fragment reassembly (SNB-0014)

+++ src/capture/reassembly.rs
@@ -33,7 +33,7 @@ const MAX_TCP_BUFFER: usize = 65536;
 struct FragmentKey {
     src: IpAddr,
     dst: IpAddr,
-    ip_id: u16,
+    ip_id: u32,
     protocol: u8,
 }
 
@@ -562,7 +562,7 @@ mod tests {
     fn make_fragment(
         src: IpAddr,
         dst: IpAddr,
-        ip_id: u16,
+        ip_id: u32,
         offset: u16, // in 8-byte units
         more_fragments: bool,
         payload: &[u8],
@@ -660,7 +660,7 @@ mod tests {
         let mut r = FragmentReassembler::with_limits(1000, DEFAULT_TTL);
         let src = IpAddr::V4(Ipv4Addr::new(10, 0, 0, 1));
         let dst = IpAddr::V4(Ipv4Addr::new(10, 0, 0, 2));
-        for i in 0..1001u16 {
+        for i in 0..1001u32 {
             // Unique ip_id per fragment → unique reassembly key.
             let f = make_fragment(src, dst, i, 0, true, &[0xAA; 8]);
             r.insert(&f);
